+++ routes/user.js
@@ -1,13 +1,10 @@
 const express = require('express')
-const { get, post, destroy } = require('../controllers/users')
-const { validateRequest } = require('../middlewares')
-const { userSchema } = require('../schemas/user')
+const { get, destroy } = require('../controllers/users')
 
 const router = new express.Router()
 
 /* GET users listing. */
 router.get('/', get)
-router.post('/auth/register', validateRequest(userSchema), post)
 router.delete('/:id', destroy)
 
 module.exports = router
