+++ services/user.js
@@ -9,31 +9,10 @@ exports.getUsers = async () => {
   }
 }
 
-exports.getUserByEmail = async (email) => {
-  try {
-    const user = await User.findOne({
-      where: { email },
-      atributtes: { include: ['email'] },
-    })
-    return user
-  } catch (err) {
-    throw Error(err.message)
-  }
-}
-
 exports.deleteUser = async (id) => {
   try {
     const deleteUser = await User.destroy({ where: { id } })
     return deleteUser
-  } catch (err) {
-    throw Error(err.message)
-  }
-}
-
-exports.addUser = async (data) => {
-  try {
-    const userCreate = await User.create(data)
-    return userCreate
   } catch (err) {
     throw Error(err.message)
   }
